\name{angle_AB}
\alias{angle_AB}
\title{Angles between subspaces}
\description{ This function computes principal angles between subspaces}
\usage{angle_AB(A, B)}

\arguments{
  \item{A}{ matrix whose columns form a basis of a subspace}
  \item{B}{ matrix whose columns form a basis of a subspace}
}

\details{The function finds the principal angles between both subspaces}
\value{
 \item{}{vector of length minimum number of columns of A and B containing the ordered principal angles (in degrees) between the subspaces generated by the columns of A and B}}
\examples{
p=10
AA=as.matrix(cbind(c(1,rep(0,p-1)),c(0,1,rep(0,p-2))))
BB=as.matrix(cbind(c(rep(0,p-1),1),c(rep(0,p-2),1,0),c(0,1,rep(0,p-2))))
angle_AB(AA,BB)
}
