% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{h}
\alias{h}
\title{computes the difference between the mean of rhoc values and 0.5}
\usage{
h(s, u, c1)
}
\arguments{
\item{s}{real positive number, the s-cale of distances}

\item{u}{vector of distances (positive numbers)}

\item{c1}{constant for rho_1 function}
}
\description{
computes the difference between the mean of rhoc values and 0.5
}
