% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_val_pfc.R
\name{initial_d0}
\alias{initial_d0}
\title{computes initial value of the coefficients, beta0, and
the S robust covariance as initial value of the covariance matrix delta=}
\usage{
initial_d0(X, aux, efficiency = 0.85)
}
\arguments{
\item{X}{data matrix}

\item{aux}{constanst for both rho functions}

\item{efficiency}{optionally the efficiency of the univariate estimate}
}
\description{
computes initial value of the coefficients, beta0, and
the S robust covariance as initial value of the covariance matrix delta=
}
