% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{MscaleNR}
\alias{MscaleNR}
\title{computes Mscale using Newton Raphson}
\usage{
MscaleNR(u, c1, k1)
}
\arguments{
\item{u}{vector of distances (positive numbers)}

\item{c1}{constant for rho_1 function}

\item{k1}{constant for rho_1 function}
}
\description{
computes Mscale using Newton Raphson
}
