% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauestimate.R
\name{phi1}
\alias{phi1}
\title{computes phi1 function, internally used to compute the weights for the algorithm}
\usage{
phi1(delta, c2, k2, ese, dis, p)
}
\arguments{
\item{delta}{positive definite p x p matrix, the estimated covariance of errors}

\item{c2}{constant for rho_2 function}

\item{k2}{constant for rho_2 function}

\item{ese}{real positive number, the s-cale of distances}

\item{dis}{vector of distances}

\item{p}{dimension of ambient space}
}
\description{
computes phi1 function, internally used to compute the weights for the algorithm
}
