% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauestimate.R
\name{sqrtm2}
\alias{sqrtm2}
\title{computes the principal square root of the matrix A}
\usage{
sqrtm2(A)
}
\arguments{
\item{A}{positive (semi) definite p x p matrix}
}
\description{
computes the principal square root of the matrix A
}
