% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_val_pfc.R
\name{tauestimate_d0}
\alias{tauestimate_d0}
\title{computes initial tau estimates without covariables (d=0) for mu (center) and delta (dispersion)}
\usage{
tauestimate_d0(X, aux, inic)
}
\arguments{
\item{X}{data matrix}

\item{aux}{constanst for both rho functions}

\item{inic}{initial beta0 and delta0 from initial_d0 routine}
}
\description{
computes initial tau estimates without covariables (d=0) for mu (center) and delta (dispersion)
}
