% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauestimate.R
\name{weights2}
\alias{weights2}
\title{computes weights for the algorithm}
\usage{
weights2(dys, s, p, c1, c2)
}
\arguments{
\item{dys}{vector of distances}

\item{s}{real positive number, the s-cale of distances}

\item{p}{dimension of ambient space}

\item{c1}{constant for rho_1 function}

\item{c2}{constant for rho_2 function}
}
\description{
computes weights for the algorithm
}
